scraper/requirements.txt
selenium==4.9.1
beautifulsoup4==4.12.2
webdriver-manager==4.0.2
lxml==5.3.0
SQLAlchemy==2.0.25
mysql-connector-python==8.2.0
pytz==2024.2
